refactor collections page to server/client split
